null handling

+++ api/app/GraphQL/Mutations/DeleteUserSkill.php
@@ -4,6 +4,7 @@ namespace App\GraphQL\Mutations;
 
 use App\Models\ExperienceSkill;
 use App\Models\UserSkill;
+use Nuwave\Lighthouse\Exceptions\ValidationException;
 
 final class DeleteUserSkill
 {
@@ -15,7 +16,11 @@ final class DeleteUserSkill
      */
     public function __invoke($_, array $args)
     {
-        $existingModel = UserSkill::find($args['id']); // will never be null as policy check precedes this
+        $existingModel = UserSkill::find($args['id']);
+
+        if ($existingModel === null) {
+            throw ValidationException::withMessages(["UserSkillNotFound"]);
+        }
 
         // soft delete all experience_skill records containing the model then soft delete the model
         ExperienceSkill::where('user_skill_id', $existingModel->id)->delete();
